v2-modern/backend/requirements.txt
fastapi
uvicorn
spotipy
python-dotenv
pydantic
pydantic-settings
python-multipart
sqlalchemy
passlib[bcrypt]
PyJWT
